Move chain packet components to packets:: module
