+++ crates/cosmos/cosmos-client-components/src/impls/packet_from_ack.rs
@@ -1,4 +1,4 @@
-use hermes_relayer_components::chain::traits::components::packet_from_write_ack_builder::PacketFromWriteAckBuilder;
+use hermes_relayer_components::chain::traits::packet::from_write_ack::PacketFromWriteAckBuilder;
 use hermes_relayer_components::chain::traits::types::ibc_events::write_ack::HasWriteAckEvent;
 use hermes_relayer_components::chain::traits::types::packet::HasIbcPacketTypes;
 use ibc_relayer_types::core::ics04_channel::events::WriteAcknowledgement;
