+++ crates/relayer/relayer-components-extra/src/components/extra/closures/relay/ack_packet_relayer.rs
@@ -1,7 +1,7 @@
 use cgp_core::{CanRaiseError, ErrorRaiser, HasComponents, HasErrorType};
-use hermes_relayer_components::chain::traits::components::packet_fields_reader::CanReadPacketFields;
 use hermes_relayer_components::chain::traits::message_builders::ack_packet::CanBuildAckPacketMessage;
 use hermes_relayer_components::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
+use hermes_relayer_components::chain::traits::packet::fields::CanReadPacketFields;
 use hermes_relayer_components::chain::traits::payload_builders::ack_packet::CanBuildAckPacketPayload;
 use hermes_relayer_components::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
 use hermes_relayer_components::chain::traits::queries::chain_status::CanQueryChainStatus;
