--- crates/relayer/relayer-components/src/chain/traits/components/mod.rs
@@ -1,2 +1,0 @@
-pub mod packet_fields_reader;
-pub mod packet_from_write_ack_builder;
