+++ crates/relayer/relayer-components/src/components/default/closures/relay/event_relayer.rs
@@ -1,7 +1,7 @@
 use cgp_core::{ErrorRaiser, HasComponents};
 
-use crate::chain::traits::components::packet_from_write_ack_builder::CanBuildPacketFromWriteAck;
 use crate::chain::traits::logs::packet::CanLogChainPacket;
+use crate::chain::traits::packet::from_write_ack::CanBuildPacketFromWriteAck;
 use crate::chain::traits::queries::counterparty_chain_id::CanQueryCounterpartyChainId;
 use crate::chain::traits::types::chain_id::HasChainId;
 use crate::chain::traits::types::ibc_events::send_packet::HasSendPacketEvent;
