+++ crates/relayer/relayer-components/src/components/default/closures/relay/packet_relayer.rs
@@ -1,6 +1,5 @@
 use cgp_core::{ErrorRaiser, HasComponents, HasErrorType};
 
-use crate::chain::traits::components::packet_fields_reader::CanReadPacketFields;
 use crate::chain::traits::logs::packet::CanLogChainPacket;
 use crate::chain::traits::message_builders::ack_packet::CanBuildAckPacketMessage;
 use crate::chain::traits::message_builders::receive_packet::CanBuildReceivePacketMessage;
@@ -8,6 +7,7 @@ use crate::chain::traits::message_builders::timeout_unordered_packet::{
     CanBuildTimeoutUnorderedPacketMessage, CanBuildTimeoutUnorderedPacketPayload,
 };
 use crate::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
+use crate::chain::traits::packet::fields::CanReadPacketFields;
 use crate::chain::traits::payload_builders::ack_packet::CanBuildAckPacketPayload;
 use crate::chain::traits::payload_builders::receive_packet::CanBuildReceivePacketPayload;
 use crate::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
