+++ crates/solomachine/solomachine-relayer/src/impls/chain/component.rs
@@ -2,12 +2,12 @@ use cgp_core::prelude::*;
 use hermes_cosmos_client_components::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
 use hermes_cosmos_client_components::impls::packet_fields::CosmosPacketFieldReader;
 use hermes_cosmos_client_components::impls::types::chain::ProvideCosmosChainTypes;
-use hermes_relayer_components::chain::traits::components::packet_fields_reader::PacketFieldsReaderComponent;
 use hermes_relayer_components::chain::traits::message_builders::channel_handshake::ChannelHandshakeMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::connection_handshake::ConnectionHandshakeMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::timeout_unordered_packet::TimeoutUnorderedPacketMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::update_client::UpdateClientMessageBuilderComponent;
+use hermes_relayer_components::chain::traits::packet::fields::PacketFieldsReaderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::channel_handshake::ChannelHandshakePayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::connection_handshake::ConnectionHandshakePayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::create_client::CreateClientPayloadBuilderComponent;
